file open issue fixed, new main class added
