--- debug.log
@@ -1,1 +1,0 @@
-[1202/120350.014:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
