+++ nbproject/private/private.properties
@@ -10,7 +10,6 @@ file.reference.en_us.jar=D:\\Java Utilities\\freetts-1.2.2-bin\\freetts-1.2\\lib
 file.reference.freetts-jsapi10.jar=D:\\Java Utilities\\freetts-1.2.2-bin\\freetts-1.2\\lib\\freetts-jsapi10.jar
 file.reference.freetts.jar=D:\\Java Utilities\\freetts-1.2.2-bin\\freetts-1.2\\lib\\freetts.jar
 file.reference.jsapi.jar=D:\\Java Utilities\\freetts-1.2.2-bin\\freetts-1.2\\lib\\jsapi.jar
-file.reference.JTattoo-1.6.13.jar=D:\\Java Utilities\\JTattoo-1.6.13.jar
 file.reference.mbrola.jar=D:\\Java Utilities\\freetts-1.2.2-bin\\freetts-1.2\\mbrola\\mbrola.jar
 javac.debug=true
 javadoc.preview=true
