+++ nbproject/private/private.xml
@@ -3,10 +3,10 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/C:/Users/Arafat%20Hossain%20Ar/Documents/NetBeansProjects/LinaTheBot/nbproject/build-impl.xml</file>
+            <file>file:/C:/Users/Arafat%20Hossain%20Ar/Documents/NetBeansProjects/LinaTheBot/src/linaUI/LinaVoice.java</file>
             <file>file:/C:/Users/Arafat%20Hossain%20Ar/Documents/NetBeansProjects/LinaTheBot/src/linathebot/MainFrame.java</file>
-            <file>file:/C:/Users/Arafat%20Hossain%20Ar/Documents/NetBeansProjects/LinaTheBot/src/linathebot/linathebot.java</file>
             <file>file:/C:/Users/Arafat%20Hossain%20Ar/Documents/NetBeansProjects/LinaTheBot/src/voice/recognizer/voiceRecognizer.java</file>
-            <file>file:/C:/Users/Arafat%20Hossain%20Ar/Documents/NetBeansProjects/LinaTheBot/src/newpackage/LinaVoice.java</file>
         </group>
     </open-files>
 </project-private>
