+++ nbproject/project.properties
@@ -38,12 +38,13 @@ file.reference.en_us.jar=D:\\freetts-1.2\\lib\\en_us.jar
 file.reference.freetts-jsapi10.jar=D:\\freetts-1.2\\lib\\freetts-jsapi10.jar
 file.reference.freetts.jar=D:\\freetts-1.2\\lib\\freetts.jar
 file.reference.jsapi.jar=D:\\freetts-1.2\\lib\\jsapi.jar
-file.reference.JTattoo-1.6.13.jar=C:\\Users\\Arafat Hossain Ar\\Desktop\\JTattoo-1.6.13.jar
 file.reference.mbrola.jar=D:\\freetts-1.2\\lib\\mbrola.jar
 file.reference.sphinx4-core-5prealpha.jar=src\\voice\\rec\\sources\\sphinx4-core-5prealpha.jar
 file.reference.sphinx4-data-5prealpha.jar=src\\voice\\rec\\sources\\sphinx4-data-5prealpha.jar
 includes=**
+jar.archive.disabled=${jnlp.enabled}
 jar.compress=false
+jar.index=${jnlp.enabled}
 javac.classpath=\
     ${libs.absolutelayout.classpath}:\
     ${file.reference.cmu_time_awb.jar}:\
@@ -56,9 +57,10 @@ javac.classpath=\
     ${file.reference.freetts.jar}:\
     ${file.reference.jsapi.jar}:\
     ${file.reference.mbrola.jar}:\
-    ${file.reference.JTattoo-1.6.13.jar}:\
     ${file.reference.sphinx4-core-5prealpha.jar}:\
-    ${file.reference.sphinx4-data-5prealpha.jar}
+    ${file.reference.sphinx4-data-5prealpha.jar}:\
+    ${libs.swing-layout.classpath}:\
+    ${libs.beans-binding.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -83,7 +85,24 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=linathebot.linathebot
+jnlp.codebase.type=no.codebase
+jnlp.descriptor=application
+jnlp.enabled=false
+jnlp.mixed.code=default
+jnlp.offline-allowed=false
+jnlp.signed=false
+jnlp.signing=
+jnlp.signing.alias=
+jnlp.signing.keystore=
+main.class=linathebot.MainFrame
+# Optional override of default Application-Library-Allowable-Codebase attribute identifying the locations where your signed RIA is expected to be found.
+manifest.custom.application.library.allowable.codebase=
+# Optional override of default Caller-Allowable-Codebase attribute identifying the domains from which JavaScript code can make calls to your RIA without security prompts.
+manifest.custom.caller.allowable.codebase=
+# Optional override of default Codebase manifest attribute, use to prevent RIAs from being repurposed
+manifest.custom.codebase=
+# Optional override of default Permissions manifest attribute (supported values: sandbox, all-permissions)
+manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
