--- src/linathebot/linathebot.java
@@ -1,30 +1,0 @@
-/*
-Author: Arafat Hossain Ar
-Date: 12/12/2020
-Project Name: Lina The Talking Bot
-Version: 1.2
- */
-package linathebot;
-
-import java.io.IOException;
-
-/**
- *
- * @author Arafat Hossain Ar
- */
-public class linathebot {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) throws IOException {
-        // TODO code application logic here
-        
-        //Main class for LINATHEBOT
-        new MainFrame().setVisible(true);
-      
-        
-
-    }
-    
-}
